+++ CHANGES.rst
@@ -6,6 +6,7 @@ LATEST
 - Added ColouredText objects to handle embedded colour codes in text for some widgets.
 - Fixed exception on reinstating NoneType signal handler.
 - Fixed float/int issue with recent builds of pywin32.
+- Fixed issue where setting options changed the selected value (even if it was still present).
 
 1.11.0
 ------